Do not convert arguments to numpy array in tset
The `Population.tset` method allows to set the parameters of each neuron in
a neuron population individually, including the "spike_times" parameter
of a spike source. `tset` expects a `valueArray` list with the parameters
for each individual neuron, where the length of `valueArray` must be equal
to the size of the population. In the case of `spike_times`, `valueArray`
is a list of lists.

Converting a list of lists to an numpy array creates a two dimensional
array, which a) may fail due to the individual `spike_times` lists having
different lengths and b) causes the original
`self.cell.shape == values.shape` test to fail because `values.shape` is a
tuple for 2d matrices.

My fix simply removes the superfluous back-and-forth conversion to a
numpy array. In the case a numpy array is given, this code still works as
intended, as the numpy array implements all interfaces required by "len"
and "zip".

I tested the fix in my code base and it seems to work fine.

Change-Id: I727feed85ca81b75056f8f0fbca7d06ad7dd11c8